Move mask function to utils
